Encapsulate map in PipelinedRequestContext
